// libs/custom-auth/src/lib/shared/generated/models/ConfigResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * onecx-extensions-bff
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ConfigResponse
 */
export interface ConfigResponse {
  /**
   * sso idp hint
   * @type {string}
   * @memberof ConfigResponse
   */
  idpHint?: string;
  /**
   *
   * @type {string}
   * @memberof ConfigResponse
   */
  url: string;
  /**
   *
   * @type {string}
   * @memberof ConfigResponse
   */
  idmId?: string;
}

/**
 * Check if a given object implements the ConfigResponse interface.
 */
export function instanceOfConfigResponse(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && 'url' in value;

  return isInstance;
}

export function ConfigResponseFromJSON(json: any): ConfigResponse {
  return ConfigResponseFromJSONTyped(json, false);
}

export function ConfigResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): ConfigResponse {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    idpHint: !exists(json, 'idpHint') ? undefined : json['idpHint'],
    url: json['url'],
    idmId: !exists(json, 'idmId') ? undefined : json['idmId'],
  };
}

export function ConfigResponseToJSON(value?: ConfigResponse | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    idpHint: value.idpHint,
    url: value.url,
    idmId: value.idmId,
  };
}